+++ CHANGELOG.md
@@ -5,7 +5,7 @@ Changelog
 
 Bug fixes:
 
-  - Do not report globals as undefined #2302
+  - Do not report globals or super globals as undefined #2302
 
 ## 2023.06.17